rename TypeReference.java to AstVisitorValueDescriptor
